requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
graphviz>=0.20.1 
boto3>=1.37.26
